flask_app/requirements.txt
Flask==3.1.2
joblib==1.5.2
mlflow==3.4.0
mlflow_skinny==3.4.0
mlflow_tracing==3.4.0
pandas==2.3.3
python-dotenv==1.1.1
contractions==0.1.73
nltk==3.9.1
pandas==2.3.3
scikit_learn==1.7.2
spacy==3.8.7
